Fix local AI tool streaming edge cases
